# openapi.yaml
openapi: 3.0.3
info:
  title: RAG Service
  version: 2.1.0
paths:
...
  /ingestion/logs:
    get:
      tags:
      - ingestion
      summary: List Logs
      operationId: list_logs_ingestion_logs_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Offset
      - name: wallet_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Wallet Id
      - name: app_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: App Id
      - name: kb_key
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Kb Key
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionLogList'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - ingestion
      summary: Create Log
      operationId: create_log_ingestion_logs_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestionLogCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /ingestion/upload:
    post:
      tags:
      - ingestion
      summary: Upload File
      operationId: upload_file_ingestion_upload_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_file_ingestion_upload_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/public/auth/challenge:
    post:
      tags:
      - auth
      summary: Auth Challenge
      operationId: auth_challenge_api_v1_public_auth_challenge_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthChallengeRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/public/auth/verify:
    post:
      tags:
      - auth
      summary: Auth Verify
      operationId: auth_verify_api_v1_public_auth_verify_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthVerifyRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AuthChallengeRequest:
      properties:
        address:
          type: string
          title: Address
          description: EVM address (0x...)
      type: object
      required:
      - address
      title: AuthChallengeRequest
    AuthVerifyRequest:
      properties:
        address:
          type: string
          title: Address
          description: EVM address (0x...)
        signature:
          type: string
          title: Signature
          description: Signature for the issued challenge (personal_sign)
      type: object
      required:
      - address
      - signature
      title: AuthVerifyRequest
    Body_upload_file_ingestion_upload_post:
      properties:
        wallet_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Wallet Id
        app_id:
          type: string
          title: App Id
        kb_key:
          type: string
          title: Kb Key
        data_wallet_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Data Wallet Id
        filename:
          anyOf:
          - type: string
          - type: 'null'
          title: Filename
        file:
          type: string
          format: binary
          title: File
      type: object
      required:
      - app_id
      - kb_key
      - file
      title: Body_upload_file_ingestion_upload_post
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    IngestionLogCreate:
      properties:
        wallet_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Wallet Id
          description: 开发者钱包 ID（兼容字段；推荐使用 Authorization）
        status:
          type: string
          title: Status
          description: started|success|failed|info
        message:
          anyOf:
          - type: string
          - type: 'null'
          title: Message
          description: Log message
          default: ''
        app_id:
          type: string
          title: App Id
          description: Plugin app_id
        kb_key:
          anyOf:
          - type: string
          - type: 'null'
          title: Kb Key
          description: KB key
        collection:
          anyOf:
          - type: string
          - type: 'null'
          title: Collection
          description: Vector collection
        meta:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Meta
      type: object
      required:
      - status
      - app_id
      title: IngestionLogCreate
    IngestionLogItem:
      properties:
        id:
          type: integer
          title: Id
        status:
          type: string
          title: Status
        message:
          anyOf:
          - type: string
          - type: 'null'
          title: Message
        wallet_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Wallet Id
        app_id:
          anyOf:
          - type: string
          - type: 'null'
          title: App Id
        kb_key:
          anyOf:
          - type: string
          - type: 'null'
          title: Kb Key
        collection:
          anyOf:
          - type: string
          - type: 'null'
          title: Collection
        meta_json:
          anyOf:
          - type: string
          - type: 'null'
          title: Meta Json
        created_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Created At
      type: object
      required:
      - id
      - status
      title: IngestionLogItem
    IngestionLogList:
      properties:
        items:
          items:
            $ref: '#/components/schemas/IngestionLogItem'
          type: array
          title: Items
      type: object
      title: IngestionLogList
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...